Added 2 new headers (fieldnames) for our CSV file: StartDate and EndDate.
Created correct_date function to gather the user's input, convert it to a date format and validate it against specific rules.

The function will return the start date and end date (if both correct) that will be used to update our CSV file.

If the start date or end date (or both) are not correct (the user has 3 tries), the function will return False and the book will not be added/updated and will circle back to the main menu.