Implemented new add book feature.
Created a function to read and write the file, used try->except->else to make sure that the file is read and written successfully.

Added conditions for book name and author name to be of at least 3 characters long.

Fixed bug where the headers were added with each entry to our CSV file.

Added conditions to check if the book already exists in our CSV file.

Linked the correct_date function (previously created) with the add book feature to gather the start and end dates for a book that was finished (also asking the user if the book was read or not).

Used conditions to ask the user if he/she wants to leave a comment or share the book, these 2 options are available only if the book was read.

Used try->except->else for user's input.