Implemented a new update book feature.
Created a reader & writer function (similar to the one for add book feature).

Used try->except->else to make sure that the CSV file is read/written without errors.

Used try->except->else for user's input.

Used conditions to make the edit feature available only if the book is read, otherwise gave the option to mark the book as read by setting the start and end dates using the correct_date function created before.

If the dates provided are incorrect, nothing will be updated and the app will return to the main menu.

Also used conditions to determine if the book is already read and inform the user that no updates will be made.

Fixed bug where only the updated book's details where added to the file, the remaining rows were deleted.

The updated book will always be listed as the first row now, this way reflecting that it's the most recent updated entry.

Linked the add_book function for the scenario where the book's name is not found in our file.