Updated the list books feature.
Used try->except->else to make sure that the file is read without errors.

Added the new fields from our CSV file to the FOR loop and listed the books as a table.

Added new feature that will check if the file is empty and will print an error message if there are no books in the CSV.

Linked this feature with the update book function, this way if the user selects update book and the file is empty a warning message will be printed.