Updated the share book feature.
Created a reader and writer function similar to the one used for add book and update book.

Used conditions to make a book shareable only if it's marked as read.

Linked the share book feature to the correct date function to gather the start and end dates.

Provided the option to share a book with multiple people.

Used try->except->else:

- for user's inputs
- to make sure that the file is read/written without any errors

Linked the feature with the updated list books and add book features.