Implemented the book notes/review feature in a new module and imported it in the main page of the app.
Used strip and title functions to make sure that the user's input is in the same format as the data from the CSV file.

Linked the date, list books and add book features.

Added conditions to make the note/review of at least 3 characters and no more than 30.

Added conditions to make the note/review possible only for a book that was read.