Implemented the book notes/review feature in a new module.
Used strip and title functions to make sure that the user's input is in the same format as the data from the CSV file.

Linked the date, list books and add book features.

Added conditions to make the note/review of at least 3 characters and no more than 30.

Added conditions to make the note/review possible only for a book that was read.

Created a tools module with a Utils class where several static methods are stored: clear screen, main menu, app start.

Also created 2 new static methods, one that will return several empty variables that are used with all the features and are taking values from the conditional tests.

Find book is a method that determines if a specified book is in our CSV file.